+++ merk/src/proofs/query/query_item/intersect.rs
@@ -1,14 +1,144 @@
-use std::ops::{Range, RangeFrom, RangeFull, RangeInclusive, RangeTo, RangeToInclusive};
+use std::{
+    cmp::Ordering,
+    ops::{Range, RangeFrom, RangeFull, RangeInclusive, RangeTo, RangeToInclusive},
+};
 
-use crate::{proofs::query::query_item::QueryItem, Op::Put};
+use crate::proofs::query::query_item::{
+    intersect::RangeSetItem::{
+        ExclusiveEnd, ExclusiveStart, Inclusive, UnboundedEnd, UnboundedStart,
+    },
+    QueryItem,
+};
+
+pub struct RangeSetIntersection {
+    in_both: Option<RangeSet>,
+    ours_left: Option<RangeSet>,
+    ours_right: Option<RangeSet>,
+    theirs_left: Option<RangeSet>,
+    theirs_right: Option<RangeSet>,
+}
+
+/// Concise query item representation
+#[derive(Clone, Debug)]
+pub struct RangeSet {
+    pub start: RangeSetItem,
+    pub end: RangeSetItem,
+}
 
 #[derive(Default)]
+pub struct QueryItemManyIntersectionResult {
+    pub in_both: Option<Vec<QueryItem>>,
+    pub ours: Option<Vec<QueryItem>>,
+    pub theirs: Option<Vec<QueryItem>>,
+}
+
+pub struct QueryItemIntersectionResultTheirsLeftovers {
+    pub theirs_left: Option<QueryItem>,
+    pub theirs_right: Option<QueryItem>,
+}
+
+impl QueryItemManyIntersectionResult {
+    fn push_ours(&mut self, our_query_item: QueryItem) {
+        let ours_vec = self.ours.get_or_insert(vec![]);
+        ours_vec.push(our_query_item);
+    }
+
+    fn push_theirs(&mut self, their_query_item: QueryItem) {
+        let theirs_vec = self.theirs.get_or_insert(vec![]);
+        theirs_vec.push(their_query_item);
+    }
+
+    fn push_ours_and_in_both_from_result(
+        &mut self,
+        query_item_intersection_result: QueryItemIntersectionResult,
+    ) -> QueryItemIntersectionResultTheirsLeftovers {
+        let QueryItemIntersectionResult {
+            in_both,
+            ours_left,
+            ours_right,
+            theirs_left,
+            theirs_right,
+        } = query_item_intersection_result;
+        if let Some(in_both) = in_both {
+            let in_both_vec = self.in_both.get_or_insert(vec![]);
+            in_both_vec.push(in_both);
+        }
+        if let Some(ours_left) = ours_left {
+            let ours_vec = self.ours.get_or_insert(vec![]);
+            ours_vec.push(ours_left);
+        }
+        if let Some(ours_right) = ours_right {
+            let ours_vec = self.ours.get_or_insert(vec![]);
+            ours_vec.push(ours_right);
+        }
+
+        QueryItemIntersectionResultTheirsLeftovers {
+            theirs_left,
+            theirs_right,
+        }
+    }
+
+    fn push_theirs_from_result(
+        &mut self,
+        query_item_intersection_result: QueryItemIntersectionResultTheirsLeftovers,
+    ) {
+        let QueryItemIntersectionResultTheirsLeftovers {
+            theirs_left,
+            theirs_right,
+        } = query_item_intersection_result;
+        if let Some(theirs_left) = theirs_left {
+            let theirs_vec = self.theirs.get_or_insert(vec![]);
+            theirs_vec.push(theirs_left);
+        }
+        if let Some(theirs_right) = theirs_right {
+            let theirs_vec = self.theirs.get_or_insert(vec![]);
+            theirs_vec.push(theirs_right);
+        }
+    }
+
+    fn merge_in(&mut self, query_item_many_intersection_result: Self) {
+        let QueryItemManyIntersectionResult {
+            in_both,
+            ours,
+            theirs,
+        } = query_item_many_intersection_result;
+        if let Some(mut in_both) = in_both {
+            let in_both_vec = self.in_both.get_or_insert(vec![]);
+            in_both_vec.append(&mut in_both);
+        }
+        if let Some(mut ours) = ours {
+            let ours_vec = self.ours.get_or_insert(vec![]);
+            ours_vec.append(&mut ours);
+        }
+        if let Some(mut theirs) = theirs {
+            let theirs_vec = self.theirs.get_or_insert(vec![]);
+            theirs_vec.append(&mut theirs);
+        }
+    }
+}
+
 pub struct QueryItemIntersectionResult {
-    in_both: Option<QueryItem>,
-    ours_left: Option<QueryItem>,
-    ours_right: Option<QueryItem>,
-    theirs_left: Option<QueryItem>,
-    theirs_right: Option<QueryItem>,
+    pub in_both: Option<QueryItem>,
+    pub ours_left: Option<QueryItem>,
+    pub ours_right: Option<QueryItem>,
+    pub theirs_left: Option<QueryItem>,
+    pub theirs_right: Option<QueryItem>,
+}
+
+impl From<RangeSetIntersection> for QueryItemIntersectionResult {
+    fn from(range_set_intersection: RangeSetIntersection) -> Self {
+        Self {
+            in_both: range_set_intersection.in_both.map(|a| a.to_query_item()),
+            ours_left: range_set_intersection.ours_left.map(|a| a.to_query_item()),
+            ours_right: range_set_intersection.ours_right.map(|a| a.to_query_item()),
+            theirs_left: range_set_intersection
+                .theirs_left
+                .map(|a| a.to_query_item()),
+            theirs_right: range_set_intersection
+                .theirs_right
+                .map(|a| a.to_query_item()),
+        }
+    }
 }
 
 impl QueryItemIntersectionResult {
@@ -23,910 +153,556 @@ impl QueryItemIntersectionResult {
     }
 }
 
+impl RangeSet {
+    // TODO: convert to impl of From/To trait
+    pub fn to_query_item(&self) -> QueryItem {
+        match (&self.start, &self.end) {
+            (RangeSetItem::Inclusive(start), RangeSetItem::Inclusive(end)) => {
+                if start == end {
+                    QueryItem::Key(start.clone())
+                } else {
+                    QueryItem::RangeInclusive(RangeInclusive::new(start.clone(), end.clone()))
+                }
+            }
+            (RangeSetItem::Inclusive(start), RangeSetItem::ExclusiveEnd(end)) => {
+                QueryItem::Range(Range {
+                    start: start.clone(),
+                    end: end.clone(),
+                })
+            }
+            (RangeSetItem::Inclusive(start), RangeSetItem::UnboundedEnd) => {
+                QueryItem::RangeFrom(RangeFrom {
+                    start: start.clone(),
+                })
+            }
+            (RangeSetItem::ExclusiveStart(start), RangeSetItem::ExclusiveEnd(end)) => {
+                QueryItem::RangeAfterTo(Range {
+                    start: start.clone(),
+                    end: end.clone(),
+                })
+            }
+            (RangeSetItem::ExclusiveStart(start), RangeSetItem::Inclusive(end)) => {
+                QueryItem::RangeAfterToInclusive(RangeInclusive::new(start.clone(), end.clone()))
+            }
+            (RangeSetItem::ExclusiveStart(start), RangeSetItem::UnboundedEnd) => {
+                QueryItem::RangeAfter(RangeFrom {
+                    start: start.clone(),
+                })
+            }
+            (RangeSetItem::UnboundedStart, RangeSetItem::UnboundedEnd) => {
+                QueryItem::RangeFull(RangeFull)
+            }
+            (RangeSetItem::UnboundedStart, RangeSetItem::Inclusive(end)) => {
+                QueryItem::RangeToInclusive(RangeToInclusive { end: end.clone() })
+            }
+            (RangeSetItem::UnboundedStart, RangeSetItem::ExclusiveEnd(end)) => {
+                QueryItem::RangeTo(RangeTo { end: end.clone() })
+            }
+            _ => {
+                // TODO: return proper error, this should be unreachable
+                //  if the range set was created from a valid query item,
+                //  actually should return None in this case
+                unreachable!()
+            }
+        }
+    }
+
+    pub fn intersect(&self, other: RangeSet) -> RangeSetIntersection {
+        // check if the range sets do not overlap
+        if self.end < other.start || other.end < self.start {
+            // the sets do not overlap
+            // no common element
+            if self.end < other.start {
+                // self is at the left
+                return RangeSetIntersection {
+                    in_both: None,
+                    ours_left: Some(self.clone()),
+                    ours_right: None,
+                    theirs_right: Some(other),
+                    theirs_left: None,
+                };
+            } else {
+                return RangeSetIntersection {
+                    in_both: None,
+                    ours_left: None,
+                    ours_right: Some(self.clone()),
+                    theirs_left: Some(other),
+                    theirs_right: None,
+                };
+            }
+        }
+
+        // sets overlap
+        let (smaller_start, bigger_start) =
+            RangeSetItem::order_items(&self.start, &other.start, self.start.cmp(&other.start));
+
+        let (smaller_end, larger_end) =
+            RangeSetItem::order_items(&self.end, &other.end, self.end.cmp(&other.end));
+
+        // assume they are equal and progressively update the common boundary
+        let mut intersection_result = RangeSetIntersection {
+            in_both: Some(self.clone()),
+            ours_left: None,
+            ours_right: None,
+            theirs_left: None,
+            theirs_right: None,
+        };
+
+        // if the comparison of the start are not equal then we have value for left
+        if self.start != other.start {
+            if &self.start == smaller_start {
+                // ours left
+                intersection_result.ours_left = Some(RangeSet {
+                    start: smaller_start.clone(),
+                    end: bigger_start.invert(false),
+                });
+            } else {
+                intersection_result.theirs_left = Some(RangeSet {
+                    start: smaller_start.clone(),
+                    end: bigger_start.invert(false),
+                });
+            }
+        }
+
+        // if the comparison of the end is not equal then we have value for right
+        if self.end != other.end {
+            if self.end > other.end {
+                // ours right
+                intersection_result.ours_right = Some(RangeSet {
+                    start: smaller_end.invert(true),
+                    end: larger_end.clone(),
+                });
+            } else {
+                intersection_result.theirs_right = Some(RangeSet {
+                    start: smaller_end.invert(true),
+                    end: larger_end.clone(),
+                });
+            }
+        }
+
+        intersection_result.in_both = Some(RangeSet {
+            start: bigger_start.clone(),
+            end: smaller_end.clone(),
+        });
+
+        intersection_result
+    }
+}
+
+/// Represents all possible value types in a range set
+#[derive(Eq, PartialEq, Clone, Debug)]
+pub enum RangeSetItem {
+    UnboundedStart,
+    UnboundedEnd,
+    Inclusive(Vec<u8>),
+    ExclusiveStart(Vec<u8>),
+    ExclusiveEnd(Vec<u8>),
+}
+
+impl RangeSetItem {
+    pub fn invert(&self, is_start: bool) -> RangeSetItem {
+        match &self {
+            RangeSetItem::Inclusive(v) => {
+                if is_start {
+                    RangeSetItem::ExclusiveStart(v.clone())
+                } else {
+                    RangeSetItem::ExclusiveEnd(v.clone())
+                }
+            }
+            RangeSetItem::ExclusiveStart(v) | RangeSetItem::ExclusiveEnd(v) => {
+                RangeSetItem::Inclusive(v.clone())
+            }
+            RangeSetItem::UnboundedStart => RangeSetItem::UnboundedStart,
+            RangeSetItem::UnboundedEnd => RangeSetItem::UnboundedEnd,
+        }
+    }
+
+    /// Given som ordering and two items, this returns the items orders
+    pub fn order_items<'a>(
+        item_one: &'a RangeSetItem,
+        item_two: &'a RangeSetItem,
+        order: Ordering,
+    ) -> (&'a RangeSetItem, &'a RangeSetItem) {
+        match order {
+            Ordering::Less => (item_one, item_two),
+            _ => (item_two, item_one),
+        }
+    }
+}
+
+impl PartialOrd for RangeSetItem {
+    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
+        Some(self.cmp(other))
+    }
+}
+
+impl Ord for RangeSetItem {
+    fn cmp(&self, other: &Self) -> Ordering {
+        match (self, other) {
+            (UnboundedStart, UnboundedStart) => Ordering::Equal,
+            (UnboundedEnd, UnboundedEnd) => Ordering::Equal,
+
+            // unbounded start begins at negative infinity so it's smaller than all other values
+            (UnboundedStart, _) => Ordering::Less,
+            (_, UnboundedStart) => Ordering::Greater,
+
+            // unbounded end stops at positive infinity so larger than all other values
+            (UnboundedEnd, _) => Ordering::Greater,
+            (_, UnboundedEnd) => Ordering::Less,
+
+            (Inclusive(v1), Inclusive(v2))
+            | (ExclusiveStart(v1), ExclusiveStart(v2))
+            | (ExclusiveEnd(v1), ExclusiveEnd(v2)) => {
+                if v1 == v2 {
+                    Ordering::Equal
+                } else if v1 < v2 {
+                    Ordering::Less
+                } else {
+                    Ordering::Greater
+                }
+            }
+
+            (Inclusive(v1), ExclusiveStart(v2)) | (ExclusiveEnd(v1), Inclusive(v2)) => {
+                if v1 <= v2 {
+                    Ordering::Less
+                } else {
+                    Ordering::Greater
+                }
+            }
+            (Inclusive(v1), ExclusiveEnd(v2)) | (ExclusiveStart(v1), Inclusive(v2)) => {
+                if v1 < v2 {
+                    Ordering::Less
+                } else {
+                    Ordering::Greater
+                }
+            }
+
+            (ExclusiveStart(v1), ExclusiveEnd(v2)) | (ExclusiveEnd(v2), ExclusiveStart(v1)) => {
+                // start goes up, end goes down
+                // if they are equal, exclusive end is smaller cause it stops just before the
+                // number
+                if v1 >= v2 {
+                    Ordering::Greater
+                } else {
+                    Ordering::Less
+                }
+            }
+        }
+    }
+}
+
 impl QueryItem {
-    pub fn intersect_with_key(&self, their_key: &Vec<u8>) -> QueryItemIntersectionResult {
-        match self {
-            QueryItem::Key(key) => {
-                if key == their_key {
-                    QueryItemIntersectionResult {
-                        in_both: Some(self.clone()),
-                        ours_left: None,
-                        ours_right: None,
-                        theirs_left: None,
-                        theirs_right: None,
-                    }
-                } else {
-                    QueryItemIntersectionResult {
-                        in_both: None,
-                        ours_left: Some(self.clone()),
-                        ours_right: None,
-                        theirs_left: Some(QueryItem::Key(their_key.clone())),
-                        theirs_right: None,
-                    }
-                }
-            }
-            QueryItem::Range(range) => {
-                if range.contains(their_key) {
-                    let ours_left = if range.start == their_key {
-                        None
-                    } else {
-                        Some(QueryItem::Range(Range {
-                            start: range.start.clone(),
-                            end: their_key.clone(),
-                        }))
-                    };
-                    let ours_right = Some(QueryItem::RangeAfterTo(Range {
-                        start: their_key.clone(),
-                        end: range.end.clone(),
-                    }));
-                    QueryItemIntersectionResult {
-                        in_both: Some(QueryItem::Key(their_key.clone())),
-                        ours_left,
-                        ours_right,
-                        theirs_left: None,
-                        theirs_right: None,
-                    }
-                } else {
-                    if their_key < range.start {
-                        QueryItemIntersectionResult {
-                            in_both: None,
-                            ours_left: Some(self.clone()),
-                            ours_right,
-                            theirs_left: None,
-                            theirs_right: Some(QueryItem::Key(their_key.clone())),
-                        }
-                    } else {
-                        QueryItemIntersectionResult {
-                            in_both: None,
-                            ours_left: Some(self.clone()),
-                            ours_right,
-                            theirs_left: None,
-                            theirs_right: Some(QueryItem::Key(their_key.clone())),
-                        }
-                    }
-                }
-            }
-            QueryItem::RangeAfterTo(range) => {
-                if range.contains(their_key) && range.start != their_key {
-                    let ours_left = Some(QueryItem::Range(Range {
-                        start: range.start.clone(),
-                        end: their_key.clone(),
-                    }));
-                    let ours_right = Some(QueryItem::RangeAfterTo(Range {
-                        start: their_key.clone(),
-                        end: range.end.clone(),
-                    }));
-                    QueryItemIntersectionResult {
-                        in_both: Some(QueryItem::Key(their_key.clone())),
-                        ours_left,
-                        ours_right,
-                        theirs_left: None,
-                        theirs_right: None,
-                    }
-                } else {
-                    QueryItemIntersectionResult {
-                        in_both: None,
-                        ours_left: Some(self.clone()),
-                        ours_right,
-                        theirs_left: Some(QueryItem::Key(their_key.clone())),
-                        theirs_right: None,
-                    }
-                }
-            }
-            QueryItem::RangeInclusive(range_inclusive) => {
-                if range_inclusive.contains(&their_key) {
-                    let ours_left = if range_inclusive.start() == &their_key {
-                        None
-                    } else {
-                        Some(QueryItem::Range(Range {
-                            start: range_inclusive.start().clone(),
-                            end: their_key.clone(),
-                        }))
-                    };
-                    let ours_right = if range_inclusive.end() == &their_key {
-                        None
-                    } else {
-                        Some(QueryItem::RangeAfterToInclusive(RangeInclusive::new(
-                            their_key.clone(),
-                            range_inclusive.end().clone(),
-                        )))
-                    };
-                    QueryItemIntersectionResult {
-                        in_both: Some(QueryItem::Key(their_key.clone())),
-                        ours_left,
-                        ours_right,
-                        theirs_left: None,
-                        theirs_right: None,
-                    }
-                } else {
-                    QueryItemIntersectionResult {
-                        in_both: None,
-                        ours_left: Some(self.clone()),
-                        ours_right,
-                        theirs_left: Some(QueryItem::Key(their_key.clone())),
-                        theirs_right: None,
-                    }
-                }
-            }
-            QueryItem::RangeAfterToInclusive(range_inclusive) => {
-                if range_inclusive.contains(their_key) && range_inclusive.start() != their_key {
-                    let ours_left = Some(QueryItem::Range(Range {
-                        start: range_inclusive.start().clone(),
-                        end: their_key.clone(),
-                    }));
-                    if range_inclusive.end() == their_key {
-                        QueryItemIntersectionResult {
-                            in_both: Some(QueryItem::Key(their_key.clone())),
-                            ours_left,
-                            ours_right: None,
-                            theirs_left: None,
-                            theirs_right: None,
-                        }
-                    } else {
-                        let ours_right = Some(QueryItem::RangeAfterToInclusive(
-                            RangeInclusive::new(their_key.clone(), range_inclusive.end().clone()),
-                        ));
-                        QueryItemIntersectionResult {
-                            in_both: Some(QueryItem::Key(their_key.clone())),
-                            ours_left,
-                            ours_right,
-                            theirs_left: None,
-                            theirs_right: None,
-                        }
-                    }
-                } else {
-                    QueryItemIntersectionResult {
-                        in_both: None,
-                        ours_left: Some(self.clone()),
-                        ours_right,
-                        theirs_left: Some(QueryItem::Key(their_key.clone())),
-                        theirs_right: None,
-                    }
-                }
-            }
-            QueryItem::RangeFull(_) => {
-                let ours_left = Some(QueryItem::RangeTo(RangeTo {
-                    end: their_key.clone(),
-                }));
-                let ours_right = Some(QueryItem::RangeAfter(RangeFrom {
-                    start: their_key.clone(),
-                }));
-                QueryItemIntersectionResult {
-                    in_both: Some(QueryItem::Key(their_key.clone())),
-                    ours_left,
-                    ours_right,
-                    theirs_left: None,
-                    theirs_right: None,
-                }
-            }
-            QueryItem::RangeFrom(range_from) => {
-                if range_from.contains(their_key) {
-                    if range_from.start == their_key {
-                        // Just remove first element, by going to a range after
-                        QueryItemIntersectionResult {
-                            in_both: Some(QueryItem::Key(their_key.clone())),
-                            ours_left: Some(QueryItem::RangeAfter(range_from.clone())),
-                            ours_right: None,
-                            theirs_left: None,
-                            theirs_right: None,
-                        }
-                    } else {
-                        let ours_left = Some(QueryItem::Range(Range {
-                            start: range_from.start.clone(),
-                            end: their_key.clone(),
-                        }));
-                        let ours_right = Some(QueryItem::RangeAfter(RangeFrom {
-                            start: their_key.clone(),
-                        }));
-                        QueryItemIntersectionResult {
-                            in_both: Some(QueryItem::Key(their_key.clone())),
-                            ours_left,
-                            ours_right,
-                            theirs_left: None,
-                            theirs_right: None,
-                        }
-                    }
-                } else {
-                    QueryItemIntersectionResult {
-                        in_both: None,
-                        ours_left: Some(self.clone()),
-                        ours_right: None,
-                        theirs_left: Some(QueryItem::Key(their_key.clone())),
-                        theirs_right: None,
-                    }
-                }
-            }
-            QueryItem::RangeAfter(range_after) => {
-                if range_after.contains(&their_key) && range_after.start != their_key {
-                    let ours_left = Some(QueryItem::RangeAfterTo(Range {
-                        start: range_from.start.clone(),
-                        end: their_key.clone(),
-                    }));
-                    let ours_right = Some(QueryItem::RangeAfter(RangeFrom {
-                        start: their_key.clone(),
-                    }));
-                    QueryItemIntersectionResult {
-                        in_both: Some(QueryItem::Key(their_key.clone())),
-                        ours_left,
-                        ours_right,
-                        theirs_left: None,
-                        theirs_right: None,
-                    }
-                } else {
-                    QueryItemIntersectionResult {
-                        in_both: None,
-                        ours_left: Some(self.clone()),
-                        ours_right: None,
-                        theirs_left: Some(QueryItem::Key(their_key.clone())),
-                        theirs_right: None,
-                    }
-                }
-            }
-            QueryItem::RangeTo(range_to) => {
-                if range_to.contains(their_key) {
-                    let ours_left = Some(QueryItem::RangeTo(RangeTo {
-                        end: their_key.clone(),
-                    }));
-                    let ours_right = Some(QueryItem::RangeAfterTo(Range {
-                        start: their_key.clone(),
-                        end: range_to.end.clone(),
-                    }));
-                    QueryItemIntersectionResult {
-                        in_both: Some(QueryItem::Key(their_key.clone())),
-                        ours_left,
-                        ours_right,
-                        theirs_left: None,
-                        theirs_right: None,
-                    }
-                } else {
-                    QueryItemIntersectionResult {
-                        in_both: None,
-                        ours_left: Some(self.clone()),
-                        ours_right: None,
-                        theirs_left: Some(QueryItem::Key(their_key.clone())),
-                        theirs_right: None,
-                    }
-                }
-            }
-            QueryItem::RangeToInclusive(range_to_inclusive) => {
-                if range_to_inclusive.contains(their_key) {
-                    if range_to.end == their_key {
-                        // Just remove first element, by going to a range after
-                        QueryItemIntersectionResult {
-                            in_both: Some(QueryItem::Key(their_key.clone())),
-                            ours_left: Some(QueryItem::RangeTo(RangeTo {
-                                end: their_key.clone(),
-                            })),
-                            ours_right: None,
-                            theirs_left: None,
-                            theirs_right: None,
-                        }
-                    } else {
-                        let ours_left = Some(QueryItem::RangeTo(RangeTo {
-                            end: their_key.clone(),
-                        }));
-                        let ours_right = Some(QueryItem::RangeAfterToInclusive(
-                            RangeInclusive::new(their_key.clone(), range_to.end.clone()),
-                        ));
-                        QueryItemIntersectionResult {
-                            in_both: Some(QueryItem::Key(their_key.clone())),
-                            ours_left,
-                            ours_right,
-                            theirs_left: None,
-                            theirs_right: None,
-                        }
-                    }
-                } else {
-                    QueryItemIntersectionResult {
-                        in_both: None,
-                        ours_left: Some(self.clone()),
-                        ours_right: None,
-                        theirs_left: Some(QueryItem::Key(their_key.clone())),
-                        theirs_right: None,
-                    }
-                }
-            }
-        }
-    }
-
-    pub fn intersect_with_range_full(&self) -> QueryItemIntersectionResult {
-        match self {
-            QueryItem::Key(our_key) => {
-                let theirs_left = Some(QueryItem::RangeTo(RangeTo {
-                    end: our_key.clone(),
-                }));
-                let theirs_right = Some(QueryItem::RangeAfter(RangeFrom {
-                    start: our_key.clone(),
-                }));
-                QueryItemIntersectionResult {
-                    in_both: Some(QueryItem::Key(their_key.clone())),
-                    ours_left: None,
-                    ours_right: None,
-                    theirs_left,
-                    theirs_right,
-                }
-            }
-            QueryItem::RangeFull(_) => QueryItemIntersectionResult {
-                in_both: Some(QueryItem::RangeFull(RangeFull)),
-                ours_left: None,
-                ours_right: None,
-                theirs_left: None,
-                theirs_right: None,
-            },
-            QueryItem::Range(range) => {
-                let theirs_left = Some(QueryItem::RangeTo(RangeTo {
-                    end: range.start.clone(),
-                }));
-                let theirs_right = Some(QueryItem::RangeFrom(RangeFrom {
-                    start: range.end.clone(),
-                }));
-                QueryItemIntersectionResult {
-                    in_both: Some(self.clone()),
-                    ours_left: None,
-                    ours_right: None,
-                    theirs_left,
-                    theirs_right,
-                }
-            }
-            QueryItem::RangeInclusive(range_inclusive) => {
-                let theirs_left = Some(QueryItem::RangeTo(RangeTo {
-                    end: range_inclusive.start().clone(),
-                }));
-                let theirs_right = Some(QueryItem::RangeAfter(RangeFrom {
-                    start: range_inclusive.end().clone(),
-                }));
-                QueryItemIntersectionResult {
-                    in_both: Some(self.clone()),
-                    ours_left: None,
-                    ours_right: None,
-                    theirs_left,
-                    theirs_right,
-                }
-            }
-            QueryItem::RangeAfterTo(range) => {
-                let theirs_left = Some(QueryItem::RangeToInclusive(RangeToInclusive {
-                    end: range.start.clone(),
-                }));
-                let theirs_right = Some(QueryItem::RangeFrom(RangeFrom {
-                    start: range.end.clone(),
-                }));
-                QueryItemIntersectionResult {
-                    in_both: Some(self.clone()),
-                    ours_left: None,
-                    ours_right: None,
-                    theirs_left,
-                    theirs_right,
-                }
-            }
-            QueryItem::RangeAfterToInclusive(range_inclusive) => {
-                let theirs_left = Some(QueryItem::RangeToInclusive(RangeToInclusive {
-                    end: range_inclusive.start().clone(),
-                }));
-                let theirs_right = Some(QueryItem::RangeAfter(RangeFrom {
-                    start: range_inclusive.end().clone(),
-                }));
-                QueryItemIntersectionResult {
-                    in_both: Some(self.clone()),
-                    ours_left: None,
-                    ours_right: None,
-                    theirs_left,
-                    theirs_right,
-                }
-            }
-            QueryItem::RangeFrom(range_from) => {
-                let theirs_left = Some(QueryItem::RangeTo(RangeTo {
-                    end: range_from.start.clone(),
-                }));
-                QueryItemIntersectionResult {
-                    in_both: Some(self.clone()),
-                    ours_left: None,
-                    ours_right: None,
-                    theirs_left,
-                    theirs_right: None,
-                }
-            }
-            QueryItem::RangeAfter(range_after) => {
-                let theirs_left = Some(QueryItem::RangeToInclusive(RangeToInclusive {
-                    end: range_after.start.clone(),
-                }));
-                QueryItemIntersectionResult {
-                    in_both: Some(self.clone()),
-                    ours_left: None,
-                    ours_right: None,
-                    theirs_left,
-                    theirs_right: None,
-                }
-            }
-            QueryItem::RangeTo(range_to) => {
-                let theirs_right = Some(QueryItem::RangeFrom(RangeFrom {
-                    start: range_to.end.clone(),
-                }));
-                QueryItemIntersectionResult {
-                    in_both: Some(self.clone()),
-                    ours_left: None,
-                    ours_right: None,
-                    theirs_left: None,
-                    theirs_right,
-                }
-            }
-            QueryItem::RangeToInclusive(range_to_inclusive) => {
-                let theirs_right = Some(QueryItem::RangeAfter(RangeFrom {
-                    start: range_to_inclusive.end.clone(),
-                }));
-                QueryItemIntersectionResult {
-                    in_both: Some(self.clone()),
-                    ours_left: None,
-                    ours_right: None,
-                    theirs_left: None,
-                    theirs_right,
-                }
-            }
-        }
-    }
-
-    pub fn intersect_with_range_to(
-        &self,
-        their_range_to: RangeTo<Vec<u8>>,
-    ) -> QueryItemIntersectionResult {
-        match self {
-            QueryItem::Key(our_key) => QueryItem::RangeTo(their_range_to)
-                .intersect_with_key(our_key)
-                .flip(),
-            QueryItem::RangeFull(_) => QueryItem::RangeTo(their_range_to)
-                .intersect_with_range_full()
-                .flip(),
-            QueryItem::Range(our_range) => {
-                if their_range_to.end <= our_range.start {
-                    // there is no overlap, their end is not inclusive
-                    QueryItemIntersectionResult {
-                        in_both: None,
-                        ours_left: None,
-                        ours_right: Some(self.clone()),
-                        theirs_left: Some(QueryItem::RangeTo(their_range_to)),
-                        theirs_right: None,
-                    }
-                } else if their_range_to.end >= our_range.end {
-                    // complete overlap for us
-                    let theirs_left = Some(QueryItem::RangeTo(RangeTo {
-                        end: our_range.start.clone(),
-                    }));
-                    let theirs_right = Some(QueryItem::Range(Range {
-                        start: their_range_to.end,
-                        end: our_range.end.clone(),
-                    }));
-                    QueryItemIntersectionResult {
-                        in_both: Some(self.clone()),
-                        ours_left: None,
-                        ours_right: None,
-                        theirs_left,
-                        theirs_right,
-                    }
-                } else {
-                    // partial overlap
-                    let in_both = Some(QueryItem::Range(Range {
-                        start: our_range.start.clone(),
-                        end: their_range_to.end.clone(),
-                    }));
-                    let theirs_left = Some(QueryItem::RangeTo(RangeTo {
-                        end: our_range.start.clone(),
-                    }));
-                    let ours_right = Some(QueryItem::Range(Range {
-                        start: their_range_to.end,
-                        end: our_range.end.clone(),
-                    }));
-                    QueryItemIntersectionResult {
-                        in_both,
-                        ours_left: None,
-                        ours_right,
-                        theirs_left,
-                        theirs_right: None,
-                    }
-                }
-            }
-            QueryItem::RangeInclusive(our_range_inclusive) => {
-                if &their_range_to.end <= our_range_inclusive.start() {
-                    // there is no overlap, their end is not inclusive
-                    QueryItemIntersectionResult {
-                        in_both: None,
-                        ours_left: None,
-                        ours_right: Some(self.clone()),
-                        theirs_left: Some(QueryItem::RangeTo(their_range_to)),
-                        theirs_right: None,
-                    }
-                } else if their_range_to.end == our_range_inclusive.end() {
-                    // complete overlap for us, except last element
-                    let in_both = Some(QueryItem::Range(Range {
-                        start: our_range_inclusive.start().clone(),
-                        end: our_range_inclusive.end().clone(),
-                    }));
-                    let theirs_left = Some(QueryItem::RangeTo(RangeTo {
-                        end: our_range.start.clone(),
-                    }));
-                    let theirs_right = Some(QueryItem::Key(their_range_to.end));
-                    QueryItemIntersectionResult {
-                        in_both,
-                        ours_left: None,
-                        ours_right: None,
-                        theirs_left,
-                        theirs_right,
-                    }
-                } else if &their_range_to.end > our_range_inclusive.end() {
-                    // complete overlap for us
-                    let theirs_left = Some(QueryItem::RangeTo(RangeTo {
-                        end: our_range.start.clone(),
-                    }));
-                    let theirs_right = Some(QueryItem::RangeAfterTo(Range {
-                        start: their_range_to.end,
-                        end: our_range.end.clone(),
-                    }));
-                    QueryItemIntersectionResult {
-                        in_both: Some(self.clone()),
-                        ours_left: None,
-                        ours_right: None,
-                        theirs_left,
-                        theirs_right,
-                    }
-                } else {
-                    // partial overlap
-                    let in_both = Some(QueryItem::Range(Range {
-                        start: our_range.start.clone(),
-                        end: their_range_to.end.clone(),
-                    }));
-                    let theirs_left = Some(QueryItem::RangeTo(RangeTo {
-                        end: our_range.start.clone(),
-                    }));
-                    let ours_right = Some(QueryItem::RangeInclusive(RangeInclusive::new(
-                        their_range_to.end,
-                        our_range.end.clone(),
-                    )));
-                    QueryItemIntersectionResult {
-                        in_both,
-                        ours_left: None,
-                        ours_right,
-                        theirs_left,
-                        theirs_right: None,
-                    }
-                }
-            }
-            QueryItem::RangeAfterTo(our_range) => {
-                if their_range_to.end <= our_range.start {
-                    // there is no overlap, their end is not inclusive
-                    QueryItemIntersectionResult {
-                        in_both: None,
-                        ours_left: None,
-                        ours_right: Some(self.clone()),
-                        theirs_left: Some(QueryItem::RangeTo(their_range_to)),
-                        theirs_right: None,
-                    }
-                } else if their_range_to.end >= our_range.end {
-                    // complete overlap for us
-                    let theirs_left = Some(QueryItem::RangeToInclusive(RangeToInclusive {
-                        end: our_range.start.clone(),
-                    }));
-                    let theirs_right = Some(QueryItem::Range(Range {
-                        start: their_range_to.end,
-                        end: our_range.end.clone(),
-                    }));
-                    QueryItemIntersectionResult {
-                        in_both: Some(self.clone()),
-                        ours_left: None,
-                        ours_right: None,
-                        theirs_left,
-                        theirs_right,
-                    }
-                } else {
-                    // partial overlap
-                    let in_both = Some(QueryItem::RangeAfterTo(Range {
-                        start: our_range.start.clone(),
-                        end: their_range_to.end.clone(),
-                    }));
-                    let theirs_left = Some(QueryItem::RangeToInclusive(RangeToInclusive {
-                        end: our_range.start.clone(),
-                    }));
-                    let ours_right = Some(QueryItem::Range(Range {
-                        start: their_range_to.end,
-                        end: our_range.end.clone(),
-                    }));
-                    QueryItemIntersectionResult {
-                        in_both,
-                        ours_left: None,
-                        ours_right,
-                        theirs_left,
-                        theirs_right: None,
-                    }
-                }
-            }
-            QueryItem::RangeAfterToInclusive(our_range_after_to_inclusive) => {
-                if &their_range_to.end <= our_range_after_to_inclusive.start() {
-                    // there is no overlap, their end is not inclusive
-                    QueryItemIntersectionResult {
-                        in_both: None,
-                        ours_left: None,
-                        ours_right: Some(self.clone()),
-                        theirs_left: Some(QueryItem::RangeTo(their_range_to)),
-                        theirs_right: None,
-                    }
-                } else if their_range_to.end == our_range_after_to_inclusive.end() {
-                    // complete overlap for us, except last element
-                    let in_both = Some(QueryItem::RangeAfterTo(Range {
-                        start: our_range_after_to_inclusive.start().clone(),
-                        end: our_range_after_to_inclusive.end().clone(),
-                    }));
-                    let theirs_left = Some(QueryItem::RangeToInclusive(RangeToInclusive {
-                        end: our_range.start.clone(),
-                    }));
-                    let theirs_right = Some(QueryItem::Key(their_range_to.end));
-                    QueryItemIntersectionResult {
-                        in_both,
-                        ours_left: None,
-                        ours_right: None,
-                        theirs_left,
-                        theirs_right,
-                    }
-                } else if &their_range_to.end > our_range_after_to_inclusive.end() {
-                    // complete overlap for us
-                    let theirs_left = Some(QueryItem::RangeTo(RangeTo {
-                        end: our_range.start.clone(),
-                    }));
-                    let theirs_right = Some(QueryItem::RangeAfterTo(Range {
-                        start: their_range_to.end,
-                        end: our_range.end.clone(),
-                    }));
-                    QueryItemIntersectionResult {
-                        in_both: Some(self.clone()),
-                        ours_left: None,
-                        ours_right: None,
-                        theirs_left,
-                        theirs_right,
-                    }
-                } else {
-                    // partial overlap
-                    let in_both = Some(QueryItem::RangeAfterTo(Range {
-                        start: our_range.start.clone(),
-                        end: their_range_to.end.clone(),
-                    }));
-                    let theirs_left = Some(QueryItem::RangeTo(RangeTo {
-                        end: our_range.start.clone(),
-                    }));
-                    let ours_right = Some(QueryItem::RangeInclusive(RangeInclusive::new(
-                        their_range_to.end,
-                        our_range.end.clone(),
-                    )));
-                    QueryItemIntersectionResult {
-                        in_both,
-                        ours_left: None,
-                        ours_right,
-                        theirs_left,
-                        theirs_right: None,
-                    }
-                }
-            }
-            QueryItem::RangeAfter(our_range_after) => {
-                // intersecting range after with range to
-                // will only overlap if a <= b
-                // a is range after, b is range to
-                // a is not inclusive, a+ is what we care about
-                // we can't have equal right as we don't know what the step size is
-                // the overlap will be from a to b, but inverted a right
-                // oh, we could also just use theirs but in a range item that doesn't care about
-                // the inclusive nature
-                if their_range_to.end <= our_range_after.start {
-                    // no overlap
-                    QueryItemIntersectionResult {
-                        in_both: None,
-                        ours_left: Some(QueryItem::RangeTo(their_range_to)),
-                        ours_right: None,
-                        theirs_left: None,
-                        theirs_right: Some(QueryItem::RangeAfter(our_range_after.to_owned())),
-                    }
-                } else {
-                    // we have an overlap,
-                    // we have a range to from left to a inclusive
-                    // then one from a non inclusive to  b non inclusive
-                    // finally one from b inclusive to right
-                    let in_both = Some(QueryItem::RangeAfterTo(Range {
-                        start: our_range_after.start.clone(),
-                        end: their_range_to.end.clone(),
-                    }));
-                    let ours_left = Some(QueryItem::RangeToInclusive(RangeToInclusive {
-                        end: our_range_after.start.clone(),
-                    }));
-                    let theirs_right = Some(QueryItem::RangeFrom(RangeFrom {
-                        start: their_range_to.end,
-                    }));
-                    QueryItemIntersectionResult {
-                        in_both,
-                        ours_left,
-                        ours_right: None,
-                        theirs_left: None,
-                        theirs_right,
-                    }
-                }
-            }
-            QueryItem::RangeFrom(our_range_from) => {
-                // intersecting range from and range to
-                // range_from has a as inclusive
-                // check for no overlap first
-                if our_range_from.start >= their_range_to.end {
-                    // no overlap
-                    QueryItemIntersectionResult {
-                        in_both: None,
-                        ours_left: Some(QueryItem::RangeTo(their_range_to)),
-                        ours_right: None,
-                        theirs_left: None,
-                        theirs_right: Some(QueryItem::RangeFrom(our_range_from.to_owned())),
-                    }
-                } else {
-                    // overlap
-                    // we have from left to non inclusive of range from start
-                    // for common, we have range_from start inclusive to range to end non inclusive
-                    // then range_to end inclusive to right
-                    let in_both = Some(QueryItem::Range(Range {
-                        start: our_range_from.start.clone(),
-                        end: their_range_to.end.clone(),
-                    }));
-                    let ours_left = Some(QueryItem::RangeTo(RangeTo {
-                        end: our_range_from.start.clone(),
-                    }));
-                    let theirs_right = Some(QueryItem::RangeFrom(RangeFrom {
-                        start: their_range_to.end.clone(),
-                    }));
-                    QueryItemIntersectionResult {
-                        in_both,
-                        ours_left,
-                        ours_right: None,
-                        theirs_left: None,
-                        theirs_right,
-                    }
-                }
-            }
-            QueryItem::RangeTo(our_range_to) => {
-                // intersecting range to with range to
-                // they both start from the same point so there should always be an intersection
-                // nothing to our left or their left as we all come from extreme left
-                // so in_both and someone's right
-                // in_both to the smaller one, then from smaller one to the bigger one
-                // in_both will be rangeTo
-                if our_range_to.end <= their_range_to.end {
-                    // our_range is the smaller one
-                    let in_both = Some(QueryItem::RangeTo(our_range_to.to_owned()));
-                    let ours_right = Some(QueryItem::Range(Range {
-                        start: our_range_to.end.clone(),
-                        end: their_range_to.end,
-                    }));
-                    QueryItemIntersectionResult {
-                        in_both,
-                        ours_left: None,
-                        ours_right,
-                        theirs_left: None,
-                        theirs_right: None,
-                    }
-                } else {
-                    // their range is the smaller one
-                    let in_both = Some(QueryItem::RangeTo(their_range_to.to_owned()));
-                    let theirs_right = Some(QueryItem::Range(Range {
-                        start: our_range_to.end.clone(),
-                        end: their_range_to.end,
-                    }));
-                    QueryItemIntersectionResult {
-                        in_both,
-                        ours_left: None,
-                        ours_right: None,
-                        theirs_left: None,
-                        theirs_right,
-                    }
-                }
-            }
-            QueryItem::RangeToInclusive(our_range_to_inclusive) => {
-                // intersecting range_to_inclusive with range_to
-                // similar to the one above, we'd just use a rangeInclusive for one
-                if our_range_to_inclusive.end < their_range_to.end {
-                    // our_range is the smaller one
-                    let in_both = Some(QueryItem::RangeToInclusive(
-                        our_range_to_inclusive.to_owned(),
-                    ));
-                    let ours_right = Some(QueryItem::RangeAfterTo(Range {
-                        start: our_range_to_inclusive.end.clone(),
-                        end: their_range_to.end,
-                    }));
-                    QueryItemIntersectionResult {
-                        in_both,
-                        ours_left: None,
-                        ours_right,
-                        theirs_left: None,
-                        theirs_right: None,
-                    }
-                } else {
-                    // their range is the smaller one
-                    let in_both = Some(QueryItem::RangeTo(their_range_to.clone()));
-                    let theirs_right = Some(QueryItem::RangeInclusive(RangeInclusive::new(
-                        their_range_to.end,
-                        our_range_to_inclusive.end.clone(),
-                    )));
-                    QueryItemIntersectionResult {
-                        in_both,
-                        ours_left: None,
-                        ours_right: None,
-                        theirs_left: None,
-                        theirs_right,
-                    }
-                }
-            }
-        }
-    }
-
-    pub fn intersect_with_range(&self, their_range: Range<Vec<u8>>) -> QueryItemIntersectionResult {
-        match self {
-            QueryItem::Key(our_key) => QueryItem::Range(their_range)
-                .intersect_with_key(our_key)
-                .flip(),
-            QueryItem::RangeFull(_) => QueryItem::Range(their_range)
-                .intersect_with_range_full()
-                .flip(),
-            QueryItem::RangeTo(our_range_to) => QueryItem::Range(their_range)
-                .intersect_with_range_to(our_range_to.clone())
-                .flip(),
-            QueryItem::Range(our_range) => {
-                // intersecting a range with another range both ends not inclusive
-                // might not overlap at all
-                if our_range.end <= their_range.start || their_range.end <= our_range.start {
-                    // no overlap, determine which is smaller to know right or left
-                    if our_range.end <= their_range.start {
-                        // our range is to the left
-                        QueryItemIntersectionResult {
-                            in_both: None,
-                            ours_left: None,
-                            ours_right: Some(QueryItem::Range(their_range)),
-                            theirs_left: Some(QueryItem::Range(our_range.to_owned())),
-                            theirs_right: None,
-                        }
-                    } else {
-                        // their range is to the left
-                        QueryItemIntersectionResult {
-                            in_both: None,
-                            ours_left: Some(QueryItem::Range(their_range)),
-                            ours_right: None,
-                            theirs_left: None,
-                            theirs_right: Some(QueryItem::Range(our_range.to_owned())),
-                        }
-                    }
-                } else {
-                    // there is an overlap
-                    // get the smaller start and construct a range (non inclusive end) to other
-                    // start for common construct from start of other (ref
-                    // above) to smaller end for right construct from the
-                    // inverse of smaller end to bigger end
-
-                    let mut intersection_result = QueryItemIntersectionResult::default();
-
-                    // the starts could be equal tho, in that case nothing on the left
-                    if our_range.start < their_range.start {
-                        our_right
-                    }
-                }
-            }
-            QueryItem::RangeInclusive(_) => {}
-            QueryItem::RangeFrom(_) => {}
-            QueryItem::RangeToInclusive(_) => {}
-            QueryItem::RangeAfter(_) => {}
-            QueryItem::RangeAfterTo(_) => {}
-            QueryItem::RangeAfterToInclusive(_) => {}
-        }
-    }
-
     pub fn intersect(&self, other: &Self) -> QueryItemIntersectionResult {
-        match other {
-            QueryItem::Key(key) => self.intersect_with_key(key),
-            QueryItem::RangeFull(_) => self.intersect_with_range_full(),
-            QueryItem::RangeTo(range_to) => self.intersect_with_range_to(range_to.clone()),
-            QueryItem::Range(range) => self.intersect_with_range(range.clone()),
-            QueryItem::RangeInclusive(_) => {}
-            QueryItem::RangeFrom(_) => {}
+        self.to_range_set().intersect(other.to_range_set()).into()
+    }
 
-            QueryItem::RangeToInclusive(_) => {}
-            QueryItem::RangeAfter(_) => {}
-            QueryItem::RangeAfterTo(_) => {}
-            QueryItem::RangeAfterToInclusive(_) => {}
+    // TODO: convert to impl of From/To trait
+    pub fn to_range_set(&self) -> RangeSet {
+        match self {
+            QueryItem::Key(start) => RangeSet {
+                start: RangeSetItem::Inclusive(start.clone()),
+                end: RangeSetItem::Inclusive(start.clone()),
+            },
+            QueryItem::Range(range) => RangeSet {
+                start: RangeSetItem::Inclusive(range.start.clone()),
+                end: RangeSetItem::ExclusiveEnd(range.end.clone()),
+            },
+            QueryItem::RangeInclusive(range) => RangeSet {
+                start: RangeSetItem::Inclusive(range.start().clone()),
+                end: RangeSetItem::Inclusive(range.end().clone()),
+            },
+            QueryItem::RangeFull(..) => RangeSet {
+                start: RangeSetItem::UnboundedStart,
+                end: RangeSetItem::UnboundedEnd,
+            },
+            QueryItem::RangeFrom(range) => RangeSet {
+                start: RangeSetItem::Inclusive(range.start.clone()),
+                end: RangeSetItem::UnboundedEnd,
+            },
+            QueryItem::RangeTo(range) => RangeSet {
+                start: RangeSetItem::UnboundedStart,
+                end: RangeSetItem::ExclusiveEnd(range.end.clone()),
+            },
+            QueryItem::RangeToInclusive(range) => RangeSet {
+                start: RangeSetItem::UnboundedStart,
+                end: RangeSetItem::Inclusive(range.end.clone()),
+            },
+            QueryItem::RangeAfter(range) => RangeSet {
+                start: RangeSetItem::ExclusiveStart(range.start.clone()),
+                end: RangeSetItem::UnboundedEnd,
+            },
+            QueryItem::RangeAfterTo(range) => RangeSet {
+                start: RangeSetItem::ExclusiveStart(range.start.clone()),
+                end: RangeSetItem::ExclusiveEnd(range.end.clone()),
+            },
+            QueryItem::RangeAfterToInclusive(range) => RangeSet {
+                start: RangeSetItem::ExclusiveStart(range.start().clone()),
+                end: RangeSetItem::Inclusive(range.end().clone()),
+            },
         }
+    }
+
+    /// For this intersection to work ours and theirs must be ordered
+    pub fn intersect_many_ordered(
+        ours: &mut Vec<Self>,
+        theirs: Vec<Self>,
+    ) -> QueryItemManyIntersectionResult {
+        let mut result = QueryItemManyIntersectionResult::default();
+        for our_item in ours.drain(..) {
+            // We create an intersection result for this one item
+            let mut one_item_pair_intersections = QueryItemManyIntersectionResult::default();
+            // We add our item
+            // In the end the item might be split up
+            one_item_pair_intersections.push_ours(our_item);
+            for their_item in theirs.clone() {
+                // We take the vector of our item
+                // It might be empty if it has already been completely consumed
+                // Meaning that all the item was inside of their items
+                if let Some(our_item_split_sections) = one_item_pair_intersections.ours.take() {
+                    let mut maybe_temp_their_item = Some(their_item);
+                    for our_partial_item in our_item_split_sections {
+                        if let Some(temp_their_item) = maybe_temp_their_item {
+                            let intersection_result = our_partial_item.intersect(&temp_their_item);
+                            // ours and in both are guaranteed to be unique
+                            let theirs_leftovers = one_item_pair_intersections
+                                .push_ours_and_in_both_from_result(intersection_result);
+                            // if we assume theirs is ordered
+                            // then we can push the left leftover
+                            if let Some(theirs_left) = theirs_leftovers.theirs_left {
+                                one_item_pair_intersections.push_theirs(theirs_left)
+                            }
+                            maybe_temp_their_item = theirs_leftovers.theirs_right
+                        } else {
+                            // there is no more of their item left
+                            // just push our partial item
+                            one_item_pair_intersections.push_ours(our_partial_item)
+                        }
+                    }
+                    // we need to add the end theirs leftovers
+                    if let Some(theirs_left) = maybe_temp_their_item {
+                        one_item_pair_intersections.push_theirs(theirs_left)
+                    }
+                } else {
+                    one_item_pair_intersections.push_theirs(their_item)
+                }
+            }
+            result.merge_in(one_item_pair_intersections)
+        }
+        result
+    }
+}
+
+#[cfg(test)]
+mod test {
+    use std::{
+        cmp::Ordering,
+        ops::{Range, RangeInclusive},
+    };
+
+    use crate::proofs::query::query_item::{
+        intersect::{RangeSetItem},
+        QueryItem,
+    };
+
+    #[test]
+    pub fn test_range_set_query_item_conversion() {
+        assert_eq!(
+            QueryItem::Key(vec![5]).to_range_set().to_query_item(),
+            QueryItem::Key(vec![5])
+        );
+        assert_eq!(
+            QueryItem::Range(Range {
+                start: vec![2],
+                end: vec![5]
+            })
+            .to_range_set()
+            .to_query_item(),
+            QueryItem::Range(Range {
+                start: vec![2],
+                end: vec![5]
+            })
+        );
+        assert_eq!(
+            QueryItem::RangeInclusive(RangeInclusive::new(vec![2], vec![5]))
+                .to_range_set()
+                .to_query_item(),
+            QueryItem::RangeInclusive(RangeInclusive::new(vec![2], vec![5]))
+        );
+        assert_eq!(
+            QueryItem::RangeFull(..).to_range_set().to_query_item(),
+            QueryItem::RangeFull(..)
+        );
+        assert_eq!(
+            QueryItem::RangeFrom(vec![5]..)
+                .to_range_set()
+                .to_query_item(),
+            QueryItem::RangeFrom(vec![5]..)
+        );
+        assert_eq!(
+            QueryItem::RangeTo(..vec![3]).to_range_set().to_query_item(),
+            QueryItem::RangeTo(..vec![3])
+        );
+        assert_eq!(
+            QueryItem::RangeToInclusive(..=vec![3])
+                .to_range_set()
+                .to_query_item(),
+            QueryItem::RangeToInclusive(..=vec![3])
+        );
+        assert_eq!(
+            QueryItem::RangeAfter(vec![4]..)
+                .to_range_set()
+                .to_query_item(),
+            QueryItem::RangeAfter(vec![4]..)
+        );
+        assert_eq!(
+            QueryItem::RangeAfterTo(vec![3]..vec![6])
+                .to_range_set()
+                .to_query_item(),
+            QueryItem::RangeAfterTo(vec![3]..vec![6])
+        );
+        assert_eq!(
+            QueryItem::RangeAfterToInclusive(vec![3]..=vec![7])
+                .to_range_set()
+                .to_query_item(),
+            QueryItem::RangeAfterToInclusive(vec![3]..=vec![7])
+        );
+    }
+
+    #[test]
+    pub fn test_range_set_item_compare() {
+        // doing a pyramid compare, to prevent repeated test
+        // if we compare A and B, we don't compare B and A further down
+
+        // test equality
+        assert_eq!(
+            RangeSetItem::Inclusive(vec![1]).cmp(&RangeSetItem::Inclusive(vec![1])),
+            Ordering::Equal
+        );
+        assert_eq!(
+            RangeSetItem::ExclusiveStart(vec![1]).cmp(&RangeSetItem::ExclusiveStart(vec![1])),
+            Ordering::Equal
+        );
+        assert_eq!(
+            RangeSetItem::ExclusiveEnd(vec![1]).cmp(&RangeSetItem::ExclusiveEnd(vec![1])),
+            Ordering::Equal
+        );
+        assert_eq!(
+            RangeSetItem::UnboundedStart.cmp(&RangeSetItem::UnboundedStart),
+            Ordering::Equal
+        );
+        assert_eq!(
+            RangeSetItem::UnboundedEnd.cmp(&RangeSetItem::UnboundedEnd),
+            Ordering::Equal
+        );
+
+        // test same item but less value
+        assert_eq!(
+            RangeSetItem::Inclusive(vec![1]).cmp(&RangeSetItem::Inclusive(vec![2])),
+            Ordering::Less
+        );
+        assert_eq!(
+            RangeSetItem::ExclusiveStart(vec![1]).cmp(&RangeSetItem::ExclusiveStart(vec![2])),
+            Ordering::Less
+        );
+        assert_eq!(
+            RangeSetItem::ExclusiveEnd(vec![1]).cmp(&RangeSetItem::ExclusiveEnd(vec![2])),
+            Ordering::Less
+        );
+
+        // test same item but greater value
+        assert_eq!(
+            RangeSetItem::Inclusive(vec![3]).cmp(&RangeSetItem::Inclusive(vec![2])),
+            Ordering::Greater
+        );
+        assert_eq!(
+            RangeSetItem::ExclusiveStart(vec![3]).cmp(&RangeSetItem::ExclusiveStart(vec![2])),
+            Ordering::Greater
+        );
+        assert_eq!(
+            RangeSetItem::ExclusiveEnd(vec![3]).cmp(&RangeSetItem::ExclusiveEnd(vec![2])),
+            Ordering::Greater
+        );
+
+        // unbounded end is greater than everything
+        // tried creating the maximum possible vector with vec![u8::MAX; isize::MAX as
+        // usize])) but got memory allocation problems
+        assert_eq!(
+            RangeSetItem::UnboundedEnd.cmp(&RangeSetItem::Inclusive(vec![u8::MAX; 1000])),
+            Ordering::Greater
+        );
+        assert_eq!(
+            RangeSetItem::UnboundedEnd.cmp(&RangeSetItem::ExclusiveStart(vec![u8::MAX; 1000])),
+            Ordering::Greater
+        );
+        assert_eq!(
+            RangeSetItem::UnboundedEnd.cmp(&RangeSetItem::ExclusiveEnd(vec![u8::MAX; 1000])),
+            Ordering::Greater
+        );
+        assert_eq!(
+            RangeSetItem::UnboundedEnd.cmp(&RangeSetItem::UnboundedStart),
+            Ordering::Greater
+        );
+
+        // unbounded start is less than everything
+        assert_eq!(
+            RangeSetItem::UnboundedStart.cmp(&RangeSetItem::Inclusive(vec![])),
+            Ordering::Less
+        );
+        assert_eq!(
+            RangeSetItem::UnboundedStart.cmp(&RangeSetItem::ExclusiveStart(vec![])),
+            Ordering::Less
+        );
+        assert_eq!(
+            RangeSetItem::UnboundedStart.cmp(&RangeSetItem::ExclusiveEnd(vec![])),
+            Ordering::Less
+        );
+
+        // test inclusive
+        // exclusive start represents value + step_size
+        // if step size is 1 and value is 1 then it starts at 2 (basically excluding 1)
+        // hence inclusive at 1 is less since 1 < 2
+        assert_eq!(
+            RangeSetItem::Inclusive(vec![1]).cmp(&RangeSetItem::ExclusiveStart(vec![1])),
+            Ordering::Less
+        );
+        assert_eq!(
+            RangeSetItem::Inclusive(vec![0]).cmp(&RangeSetItem::ExclusiveStart(vec![1])),
+            Ordering::Less
+        );
+        assert_eq!(
+            RangeSetItem::Inclusive(vec![2]).cmp(&RangeSetItem::ExclusiveStart(vec![1])),
+            Ordering::Greater
+        );
+        // exclusive end represents value - step_size
+        // if step size is 1 and value is 1 then it represents at 0 (includes everything
+        // before 1) hence inclusive at 1 is greater since 1 > 0
+        assert_eq!(
+            RangeSetItem::Inclusive(vec![1]).cmp(&RangeSetItem::ExclusiveEnd(vec![1])),
+            Ordering::Greater
+        );
+        assert_eq!(
+            RangeSetItem::Inclusive(vec![0]).cmp(&RangeSetItem::ExclusiveEnd(vec![1])),
+            Ordering::Less
+        );
+        assert_eq!(
+            RangeSetItem::Inclusive(vec![2]).cmp(&RangeSetItem::ExclusiveEnd(vec![1])),
+            Ordering::Greater
+        );
+
+        // test exclusive start
+        // exclusive start is greater than exclusive end for >= same value
+        assert_eq!(
+            RangeSetItem::ExclusiveStart(vec![1]).cmp(&RangeSetItem::ExclusiveEnd(vec![1])),
+            Ordering::Greater
+        );
+        assert_eq!(
+            RangeSetItem::ExclusiveStart(vec![2]).cmp(&RangeSetItem::ExclusiveEnd(vec![1])),
+            Ordering::Greater
+        );
+        // but less when the value is less
+        assert_eq!(
+            RangeSetItem::ExclusiveStart(vec![1]).cmp(&RangeSetItem::ExclusiveEnd(vec![2])),
+            Ordering::Less
+        );
     }
 }
